+++ bittensor/_dataloaders/genesis_dataloader_impl.py
@@ -1,43 +1,125 @@
+# The MIT License (MIT)
+# Copyright © 2021 Yuma Rao
+
+# Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated 
+# documentation files (the “Software”), to deal in the Software without restriction, including without limitation 
+# the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, 
+# and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+
+# The above copyright notice and this permission notice shall be included in all copies or substantial portions of 
+# the Software.
+
+# THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
+# THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL 
+# THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION 
+# OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER 
+# DEALINGS IN THE SOFTWARE.
+
 import bittensor
 import torch
 import random
+
 import requests
+from requests.adapters import HTTPAdapter
+from requests.packages.urllib3.util.retry import Retry
 
 from loguru import logger
-from bittensor.dataloaders.dataloader import BittensorDataLoader
 from torch.utils.data.dataloader import DataLoader
 from torch.utils.data import Subset
 
 from loguru import logger
 logger = logger.opt(colors=True)
 
-class GenesisTextDataloader(BittensorDataLoader):
-    
-    def __init__(
-            self,
-            batch_size, 
-            block_size,
-            config = None
-        ):
-        super(GenesisTextDataloader, self).__init__()
-        
-        assert batch_size > 0, 'Batch size must be larger than 0'
-        assert block_size > 0, 'Block size must be larger than 0'
-        
-        if config == None:
-            config = BittensorDataLoader.default_config()
+class BittensorDataLoader():
+    def __init__(self):
+        # IPFS hash of the genesis dataset
+        # TODO (shibshib): Find a proper way to set this as config instead of hardcoding it.
+        # More dataset hashes can be added as we add directories for other modalities.
+        self.genesis_text_dataset_hash = "QmXwfPoh2QFYqC6cYcW8kzyd9ruFfhnUi2kVBkdhawjUzj"
 
+        # Used to retrieve directory contentx
+        self.dag_get = 'https://ipfs.infura.io:5001/api/v0/dag/get'
+        # Used to retrieve file contents
+        self.file_cat = 'https://ipfs.infura.io:5001/api/v0/cat'
+
+        # Used when current corpus has been exhausted
+        self.refresh_corpus = False
+
+    @staticmethod
+    def requests_retry_session(
+            retries=3,
+            backoff_factor=0.3,
+            status_forcelist=(500, 502, 504),
+            session=None,
+        ):
+        """ Creates a retriable session for request calls. This enables 
+        automatic retries and back-off retries should any request calls fail. 
+
+        Args:
+            retries (int, optional): Maximum number of retries. Defaults to 3.
+            backoff_factor (float, optional): Factor by which to back off if a retry fails. Defaults to 0.3.
+            status_forcelist (tuple, optional): A set of integer HTTP status codes that we should force a retry on. Defaults to (500, 502, 504).
+            session ([type], optional): Session for which to set up the retries. Defaults to None.
+
+        Returns:
+            requests.Session(): A Requests Session object set up for retries and backoff. 
+        """
+
+        session = session or requests.Session()
+        retry = Retry(
+            total=retries,
+            read=retries,
+            connect=retries,
+            backoff_factor=backoff_factor,
+            status_forcelist=status_forcelist,
+        )
+        adapter = HTTPAdapter(max_retries=retry)
+        session.mount('http://', adapter)
+        session.mount('https://', adapter)
+        return session
+    
+    def retrieve_directory(self, dir_hash: str):
+        """Connects to Infura IPFS gateway and retrieves the directory of 
+        genesis datasets.
+
+        Returns:
+            dict: A dictionary of the files inside of the genesis_datasets and their hashes.
+        """
+        session = requests.Session()
+        params = (('arg', dir_hash),)
+        session.params.update(params)
+        directory = None
+
+        response = BittensorDataLoader.requests_retry_session(session=session).post(self.dag_get)
+
+        if response.status_code == 200:
+            directory = response.json()
+        
+        return directory
+    
+    
+    def __len__(self):
+        """ Returns length of the dataset that the dataloader is processing
+        """
+        pass
+
+    def __getitem__(self, idx):
+        """returns the next batch from the dataset.
+        """
+        pass
+
+class GenesisTextDataloader( BittensorDataLoader ):
+    
+    def __init__( self, config ):
+        super(GenesisTextDataloader, self).__init__()
         self.config = config
-        self.block_size = block_size
         self.tokenizer = bittensor.__tokenizer__
-        self.batch_size = batch_size
         
         # Retrieve a random slice of the genesis dataset
         self.data = self.construct_text_corpus()
 
         # Used to refresh corpus if we've exhausted the whole dataset
         self.refresh_corpus = False
-        
     
     def retrieve_text_file(self, file_hash: str):
         """Connects to Infura IPFS gateway and retrieves the contents of 
@@ -116,7 +198,7 @@ class GenesisTextDataloader(BittensorDataLoader):
         """
 
         # If we've exhausted the dataset, retrieve another corpus.
-        if self.refresh_corpus or len(self.data) < self.block_size:
+        if self.refresh_corpus or len(self.data) < self.config.dataloader.block_size:
             self.data = self.construct_text_corpus()
             self.refresh_corpus = False
 
@@ -125,7 +207,7 @@ class GenesisTextDataloader(BittensorDataLoader):
         if epoch_length and epoch_length < len(self.data):
             
             # Set up upper bound of indices to fit the batch size we want. 
-            idx_bound = epoch_length * self.batch_size 
+            idx_bound = epoch_length * self.config.dataloader.batch_size 
             if idx_bound < len(self):
                 # Collect enough random indices to batch together using batch_size into epoch_length batches
                 random_start = random.randint(0, len(self) - idx_bound)
@@ -144,13 +226,13 @@ class GenesisTextDataloader(BittensorDataLoader):
                 self.refresh_corpus = True
                 return DataLoader(self,
                             shuffle=True,
-                            batch_size=self.batch_size,
+                            batch_size=self.config.dataloader.batch_size,
                             num_workers=self.config.dataloader.num_workers)
 
 
             # Set up dataloader
             return DataLoader(subset,
-                            batch_size=self.batch_size,
+                            batch_size=self.config.dataloader.batch_size,
                             num_workers=self.config.dataloader.num_workers)
         
         # If epoch_length is not set or it is higher than the total size of the dataset,
@@ -158,7 +240,7 @@ class GenesisTextDataloader(BittensorDataLoader):
         self.refresh_corpus = True
         return DataLoader(self,
                             shuffle=True,
-                            batch_size=self.batch_size,
+                            batch_size=self.config.dataloader.batch_size,
                             num_workers=self.config.dataloader.num_workers)
 
     def __len__(self):
@@ -167,7 +249,7 @@ class GenesisTextDataloader(BittensorDataLoader):
         Returns:
             int: length of dataset minus block size
         """
-        return len(self.data) - self.block_size
+        return len(self.data) - self.config.dataloader.block_size
 
     def __getitem__(self, idx):
         """ Returns a batch of sentences from text dataset.
@@ -178,14 +260,14 @@ class GenesisTextDataloader(BittensorDataLoader):
             Returns:
                 x
         """
-        chunk = self.data[idx:idx + self.block_size]
+        chunk = self.data[idx:idx + self.config.dataloader.block_size]
 
         dix = []
         block_num=0
-        while block_num < self.block_size and len(chunk) > block_num:
+        while block_num < self.config.dataloader.block_size and len(chunk) > block_num:
             tokenized = self.tokenizer(chunk[block_num], padding=True, truncation=True)['input_ids']
             for t in tokenized:
-                if block_num < self.block_size:
+                if block_num < self.config.dataloader.block_size:
                     dix.append(t)
                     block_num += 1
